[Refactor] 알림 토큰 저장 로직 분리
